+++ fade-text-components/src/index.js
@@ -1,7 +1,0 @@
-import FadeText from './components/fade-text'
-import scrollManager from './utils/scroll-manager'
-
-export default {
-  component: FadeText,
-  scrollManager,
-}