+++ fade-text/index.js
@@ -1,0 +1,7 @@
+import FadeText from './src/components/fade-text'
+import scrollManager from './src/utils/scroll-manager'
+
+export default {
+  component: FadeText,
+  scrollManager,
+}