+++ listing-components/src/index.js
@@ -1,7 +1,0 @@
-import TopicsList from './components/topics'
-import List from './components/list'
-
-export default {
-  TopicsList,
-  List,
-}