+++ pagination/index.js
@@ -1,0 +1,3 @@
+import Pagination from './src/components/pagination'
+
+export default Pagination